fix: refactor file and variable name
